DDF-2893 Refactor response filter to support cache-control
