+++ distribution/docs/src/main/resources/_contents/_platform-contents/managing-platform-contents.adoc
@@ -38,4 +38,6 @@ include::{adoc-include}/_tables/conf-ddf.platform.scheduler.Command-table-conten
 
 include::{adoc-include}/_tables/conf-ddf.platform.ui.config-table-contents.adoc[]
 
+include::{adoc-include}/_tables/conf-org.codice.ddf.platform.response.filter.ResponseHeaderConfig-table-contents.adoc[]
+
 |===
