+++ distribution/docs/src/main/resources/_contents/_security-contents/managing-security-contents.adoc
@@ -49,7 +49,6 @@ include::{adoc-include}/_tables/conf-ddf.security.sts.wss.configuration-table-co
 include::{adoc-include}/_tables/conf-ddf.security.pdp.realm.AuthzRealm-table-contents.adoc[]
 include::{adoc-include}/_tables/conf-org.codice.ddf.security.filter.login.Session-table-contents.adoc[]
 include::{adoc-include}/_tables/conf-org.codice.ddf.security.policy.context.impl.PolicyManager-table-contents.adoc[]
-include::{adoc-include}/_tables/conf-org.codice.ddf.security.response.filter.ResponseHeaderConfig-table-contents.adoc[]
 
 |===
