+++ distribution/docs/src/main/resources/_contents/_tables/conf-org.codice.ddf.platform.response.filter.ResponseHeaderConfig-table-contents.adoc
@@ -1,0 +1,3 @@
+|HTTP Response Security
+|org.codice.ddf.security.response.filter.ResponseHeaderConfig
+|Instructions for the client browser detailing which location and/or which type of resources may be loaded.
