+++ distribution/docs/src/main/resources/_contents/_tables/org.codice.ddf.platform.response.filter.ResponseHeaderConfig.adoc
@@ -1,0 +1,25 @@
+.[[org.codice.ddf.security.response.filter.ResponseHeaderConfig]]HTTP Response Header Configuration
+[cols="1,1m,1,3,1,1" options="header"]
+|===
+
+|Name
+|Id
+|Type
+|Description
+|Default Value
+|Required
+
+|HTTP Response Headers
+|headers
+|String cardinality=100
+|*X-Content-Security-Policy* contains instructions for the client browser detailing which location and/or which type of resources may be loaded. +
+*X-Frame-Options* can be used to indicate whether or not a browser may render a page in a frame, iframe or object. +
+ *X-XSS-Protection* stops pages from loading when they detect reflected cross-site scripting (XSS) attacks. +
+ *Cache-Control* specifies the behavior for browsers and proxies to cache web resources.
+|X-Content-Security-Policy=default-src 'none'; connect-src 'self'; script-src 'self'; style-src 'self'; img-src 'self', +
+X-Frame-Options=SAMEORIGIN, +
+X_XSS_PROTECTION=1; mode=block, +
+Cache-Control=private, max-age=604800, immutable
+|true
+
+|===
