--- distribution/docs/src/main/resources/_contents/_tables/org.codice.ddf.security.response.filter.ResponseHeaderConfig.adoc
@@ -1,34 +1,0 @@
-.[[org.codice.ddf.security.response.filter.ResponseHeaderConfig]]HTTP Response Security Configuration
-[cols="1,1m,1,3,1,1" options="header"]
-|===
-
-|Name
-|Id
-|Type
-|Description
-|Default Value
-|Required
-
-|Content Security Policy
-|xContentSecurityPolicy
-|String
-|Instructions for the client browser detailing which location and/or which type of resources may be loaded.
-|default-src 'none'; connect-src 'self'; script-src 'self'; style-src 'self'; img-src 'self'
-|true
-
-
-|X-Frame-Options
-|xFrameOptions
-|String
-|The X-Frame-Options HTTP response header can be used to indicate whether or not a browser may render a page in a frame, iframe or object.
-|SAMEORIGIN
-|true
-
-|X-XSS-Protection
-|xXssProtection
-|String
-|The HTTP X-XSS-Protection response header is a feature that stops pages from loading when they detect reflected cross-site scripting (XSS) attacks.
-|1; mode=block
-|true
-
-|===
