+++ platform/platform-app/pom.xml
@@ -13,7 +13,9 @@
  **/
 
  -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <artifactId>platform</artifactId>
@@ -290,6 +292,11 @@
             <artifactId>failsafe</artifactId>
             <version>${jodah-failsafe.version}</version>
         </dependency>
+        <dependency>
+            <groupId>ddf.platform</groupId>
+            <artifactId>platform-filter-response</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
     <properties>
         <!-- CXF Properties used in used in platform-app's features.xml
@@ -426,13 +433,13 @@
         <cxf.tika.version>1.7</cxf.tika.version>
         <cxf.jexl.version>2.1.1</cxf.jexl.version>
         <cxf.htrace.version>4.0.1-incubating</cxf.htrace.version>
-        <cxf.checkstyle.extension />
-        <cxf.jaxb.context.class />
+        <cxf.checkstyle.extension/>
+        <cxf.jaxb.context.class/>
         <cxf.jaxb.context.class.property>none</cxf.jaxb.context.class.property>
         <cxf.spring.validation.mode>VALIDATION_AUTO</cxf.spring.validation.mode>
         <cxf.downloadSources>true</cxf.downloadSources>
-        <cxf.pmd.eclipse.ruleset.ext />
-        <cxf.resources.base.path />
+        <cxf.pmd.eclipse.ruleset.ext/>
+        <cxf.resources.base.path/>
         <!-- End apache-cxf-3.1.4-src/parent/pom.xml-->
         <!-- Copied from apache-cxf-3.0.4-src/pom.xml -->
         <cxf.build-utils.version>3.1.0</cxf.build-utils.version>
@@ -483,22 +490,30 @@
                         <configuration>
                             <artifacts>
                                 <artifact>
-                                    <file>src/main/resources/DDF_Custom_Mime_Type_Resolver.csw.config</file>
+                                    <file>
+                                        src/main/resources/DDF_Custom_Mime_Type_Resolver.csw.config
+                                    </file>
                                     <type>config</type>
                                     <classifier>csw</classifier>
                                 </artifact>
                                 <artifact>
-                                    <file>src/main/resources/DDF_Custom_Mime_Type_Resolver.geojson.config</file>
+                                    <file>
+                                        src/main/resources/DDF_Custom_Mime_Type_Resolver.geojson.config
+                                    </file>
                                     <type>config</type>
                                     <classifier>geojson</classifier>
                                 </artifact>
                                 <artifact>
-                                    <file>src/main/resources/DDF_Custom_Mime_Type_Resolver.nitf.config</file>
+                                    <file>
+                                        src/main/resources/DDF_Custom_Mime_Type_Resolver.nitf.config
+                                    </file>
                                     <type>config</type>
                                     <classifier>nitf</classifier>
                                 </artifact>
                                 <artifact>
-                                    <file>src/main/resources/DDF_Custom_Mime_Type_Resolver.xml.config</file>
+                                    <file>
+                                        src/main/resources/DDF_Custom_Mime_Type_Resolver.xml.config
+                                    </file>
                                     <type>config</type>
                                     <classifier>xml</classifier>
                                 </artifact>
