+++ platform/platform-filter-response/src/main/resources/OSGI-INF/metatype/metatype.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+/**
+ * Copyright (c) Codice Foundation
+ *
+ * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either
+ * version 3 of the License, or any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public License is distributed along with this program and can be found at
+ * <http://www.gnu.org/licenses/lgpl.html>.
+ *
+ **/
+
+ -->
+<metatype:MetaData xmlns:metatype="http://www.osgi.org/xmlns/metatype/v1.0.0">
+
+    <OCD name="HTTP Response Headers"
+         id="org.codice.ddf.platform.response.filter.ResponseHeaderConfig">
+        <AD description="HTTP Headers to be applied to all outgoing responses. Format is 'header=value'."
+            name="HTTP Response Headers" id="headers" required="true" type="String"
+            cardinality="100"
+            default="X-Content-Security-Policy=default-src 'none'; connect-src 'self'; script-src 'self'; style-src 'self'; img-src 'self',X-Frame-Options=SAMEORIGIN,X_XSS_PROTECTION=1; mode=block,Cache-Control=private\, max-age=604800\, immutable"/>
+    </OCD>
+
+    <Designate pid="org.codice.ddf.platform.response.filter.ResponseHeaderConfig">
+        <Object ocdref="org.codice.ddf.platform.response.filter.ResponseHeaderConfig"/>
+    </Designate>
+
+</metatype:MetaData>
