+++ platform/pom.xml
@@ -375,5 +375,6 @@
         <module>country-converter</module>
         <module>platform-app</module>
         <module>platform-filter-clientinfo</module>
+        <module>platform-filter-response</module>
     </modules>
 </project>
