--- platform/security/filter/security-filter-response/src/main/resources/OSGI-INF/metatype/metatype.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-/**
- * Copyright (c) Codice Foundation
- *
- * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either
- * version 3 of the License, or any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public License is distributed along with this program and can be found at
- * <http://www.gnu.org/licenses/lgpl.html>.
- *
- **/
-
- -->
-<metatype:MetaData xmlns:metatype="http://www.osgi.org/xmlns/metatype/v1.0.0">
-
-    <OCD name="HTTP Response Security" id="org.codice.ddf.security.response.filter.ResponseHeaderConfig">
-        <AD description="Instructions for the client browser detailing which location and/or which type of resources may be loaded."
-            name="Content Security Policy" id="xContentSecurityPolicy" required="true" type="String"
-            default="default-src 'none'; connect-src 'self'; script-src 'self'; style-src 'self'; img-src 'self'"/>
-        <AD description="The X-Frame-Options HTTP response header can be used to indicate whether or not a browser may render a page in a frame, iframe or object."
-            name="X-Frame-Options" id="xFrameOptions" required="true" type="String"
-            default="SAMEORIGIN"/>
-        <AD description="The HTTP X-XSS-Protection response header is a feature that stops pages from loading when they detect reflected cross-site scripting (XSS) attacks."
-            name="X-XSS-Protection" id="xXssProtection" required="true" type="String"
-            default="1; mode=block"/>
-    </OCD>
-
-    <Designate pid="org.codice.ddf.security.response.filter.ResponseHeaderConfig">
-        <Object ocdref="org.codice.ddf.security.response.filter.ResponseHeaderConfig"/>
-    </Designate>
-
-</metatype:MetaData>
